move argparser as function into main
